Rename gke-diagnoser to check-gke-ingress
